pyvirg/requirements.txt
build==0.7.0
pytest==6.2.5

ipython==7.30.1
numpy==1.21.5
matplotlib==3.4.2
PyQt5==5.15.6

pandas==1.3.5
seaborn==0.11.2

pyyaml==5.4.1

-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.1+cu113
torchvision==0.11.2+cu113
torchaudio==0.10.1+cu113

